+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added a `Dictionary.remove_none()` to clean keys with `None` values ([#34](https://github.com/XaviArnaus/pyxavi/pull/34))
 - New `MastodonPublisher` class to help on publishing to Mastodon APIs ([#38](https://github.com/XaviArnaus/pyxavi/pull/38))
 - Added a logger into the `Firefish` API wrapper ([#38](https://github.com/XaviArnaus/pyxavi/pull/38))
+- The `Queue` class now is able to work stateless ([#43](https://github.com/XaviArnaus/pyxavi/pull/43))
 
 ### Changed